# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Provides access to CHD activity data by category and dossier.",
        "title": "CHD Activities API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/activities": {
            "get": {
                "description": "Returns a list of all activities",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Get all activities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Activity"
                            }
                        }
                    }
                }
            }
        },
        "/activities/category/{category}": {
            "get": {
                "description": "Returns activities for a given category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Get activities by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Activity"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/activities/dossier/{dossier}": {
            "get": {
                "description": "Returns activities for a given dossier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Get activities by dossier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dossier",
                        "name": "dossier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.DossierActivities"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "description": "Get all unique categories with activity count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/dossiers": {
            "get": {
                "description": "Get all unique dossiers with activity count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dossiers"
                ],
                "summary": "List dossiers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Activity": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "category_pretty": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "main.DossierActivities": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Activity"
                    }
                },
                "dossier_authors": {
                    "type": "string"
                },
                "dossier_content": {
                    "type": "string"
                },
                "dossier_id": {
                    "type": "string"
                },
                "dossier_name": {
                    "type": "string"
                },
                "dossier_status": {
                    "type": "string"
                }
            }
        }
    }
}